+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+- Update Typer dependency to 0.1.0 and remove click-completion dependency
+- Add a `--version` option to print the version of openapi-python-client and exit
+
 ## 0.1.2 - 2020-03-16
 - Improve handling of optional properties in generated `to_dict` function for models
 - Add PEP 561 marker file (py.typed) to generated packages